pyproject.toml
[build-system]
requires = ["setuptools>=80.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jianying-notification"
version = "0.1.0"
description = "A message queue notification Python project using Celery workers to handle video render status updates"
readme = "README.md"
requires-python = ">=3.14"
license = {text = "MIT"}
authors = [
    {name = "BlueFocusContentUniverse"}
]
dependencies = [
    "celery==5.3.4",
    "redis==5.0.1",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "flower==2.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.setuptools]
packages = ["app"]
